refactor(stylex-transform): move the binary expression's evaluation to its node
The number path and the string path for binary expressions lived in the
AST-convertor module, though the node named for the expression is what
calls them. Move both there, along with the two-sided helper bolted onto
the addition arm and the unit tests that cover the operators real StyleX
source cannot reach.

Pure move: every function body is unchanged and the test-name set is
identical before and after, so nothing an author can observe shifts.

Two consequences worth recording:

- `ast::convertors` is now a consumer of the evaluator rather than only
  its supplier. Its `expr_to_num` and `ident_to_number` each recurse into
  the number path from an `Expr::Bin` arm, so that one function is
  re-exported `pub(crate)` from `js::evaluate` — the shape the module
  already uses for `evaluate_cached` and `deopt`. The string path and the
  two-sided helper are private to the node.
- The evaluator's bridging coercion helpers are now reachable from the
  moved code without widening their module-local visibility, which is
  what the following work needs.
